apps/serverside/app/requirements.txt
Flask==2.1.1
Werkzeug==2.0.2  # Use a version that is compatible with Flask 2.x
psycopg2-binary==2.9.1
